% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit the documentation
of the parameters from.}
\usage{
default_params_doc(haplotype, ic50_prediction_tool, peptide_length, percentile)
}
\arguments{
\item{haplotype}{the MHC haplotype name}

\item{ic50_prediction_tool}{the IC50 prediction tool}

\item{peptide_length}{length of the peptide}

\item{percentile}{percentile lowest IC50s to classify an IC50
as a binder}
}
\description{
This function does nothing. It is intended to inherit the documentation
of the parameters from.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
