% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tmh_binder.R
\name{find_tmh_binder}
\alias{find_tmh_binder}
\title{Find a TMH sentence that has at least one binder}
\usage{
find_tmh_binder(haplotype, peptide_length, percentile)
}
\description{
Find a TMH sentence that has at least one binder
}
